requirements.txt
Flask
line-bot-sdk
requests
pandas
matplotlib
Pillow
cloudinary
gunicorn